+++ CHANGELOG.md
@@ -47,6 +47,12 @@ shell, and the option for a persistent bottom bar that can display realtime stat
         - **max_column_completion_results**: (int) the maximum number of completion results to
           display in a single column
 
+## 3.1.3 (TBD)
+
+- Bug Fixes
+    - Fixed issue where `delimiter_complete()` could cause more matches than display matches
+    - Fixed issue where `CommandSet` registration did not respect disabled categories
+
 ## 3.1.2 (January 26, 2026)
 
 - Bug Fixes